ToiletApiNew/requirements.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
et-xmlfile==1.1.0
MarkupPy==1.14
mercantile==1.2.1
odfpy==1.4.1
openpyxl==3.0.7
psycopg2
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.0.4